fix: address operator hint
